sample_django_ninja/pyproject.toml
[tool.poetry]
name = "sample-django-ninja"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.6"
psycopg2 = "^2.9.9"
pillow = "^10.3.0"
pydantic = "^2.7.3"
django-ninja = "^1.1.0"
nanoid = "^2.0.0"
python-magic = "^0.4.27"

[tool.poetry.group.dev.dependencies]
rich = "^13.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
